♻️ Move `imaVideo.js` so it can be owned by Bento and Components
