+++ extensions/amp-ima-video/0.1/amp-ima-video.js
@@ -15,7 +15,7 @@
  */
 
 import {Deferred} from '../../../src/core/data-structures/promise';
-import {ImaPlayerData} from '../../../ads/google/ima-player-data';
+import {ImaPlayerData} from '../../../ads/google/ima/ima-player-data';
 import {PauseHelper} from '../../../src/utils/pause-helper';
 import {Services} from '../../../src/services';
 import {VideoEvents} from '../../../src/video-interface';
